src/java/model/InventoryItem.java
package model;

import java.sql.Date;

public class InventoryItem {
    private int itemId;
    private String itemName;
    private int quantity;
    private Date expirationDate;
    private boolean isSurplus;

    // Constructor, getters and setters
    public InventoryItem() {
    }

    public int getItemId() {
        return itemId;
    }

    public void setItemId(int itemId) {
        this.itemId = itemId;
    }

    public String getItemName() {
        return itemName;
    }

    public void setItemName(String itemName) {
        this.itemName = itemName;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }

    public Date getExpirationDate() {
        return expirationDate;
    }

    public void setExpirationDate(Date expirationDate) {
        this.expirationDate = expirationDate;
    }

    public boolean isSurplus() {
        return isSurplus;
    }

    public void setSurplus(boolean surplus) {
        this.isSurplus = surplus;
    }
}
